+++ RELEASE_NOTES.md
@@ -3,8 +3,10 @@
 Stats rework
 
 - Added new stats
+- Add graphs to the stats for an action, allowing you to compare up to 3 actions at a time
+- Allow a user to change the label of a statistic ('total count' to 'times done', for example)
+- Set a default order for stats (can be overridden by the user)
 - Truncation of long stats
 - Clarity between unique days and total counts
-- Highlighting appropriate stats value
-- Set a default order for stats (can be overridden by the user)
-- Allow a user to change the label of a statistic ('total count' to 'times done', for example)
+- Proper highlighting of stats value
+- Consistent font sizes for the stats cards